+++ servant/CHANGELOG.md
@@ -1,5 +1,7 @@
 next
 ----
+* BACKWARDS INCOMPATIBLE replace use of `ToFromByteString` with `To/FromHttpApiData` for `GetHeaders/BuildHeadersTo`
+* Add Servant.API.Times for parsing times with a format specified in their type.
 * Added Eq, Show, Read, Generic and Ord instances to IsSecure
 * BACKWARDS INCOMPATIBLE replace use of `ToFromByteString` with `To/FromHttpApiData` for `GetHeaders/BuildHeadersTo`
 * BACKWARD INCOMPATIBLE: Moved `From/ToFormUrlEncoded` classes, which were renamed to `From/ToForm` to `http-api-data`