+++ servant/CHANGELOG.md
@@ -1,5 +1,5 @@
-upcoming
---------
+0.9.1
+------
 
 * Added 'noHeader' function for *not* adding response headers.